+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # doozy CHANGELOG
 
+## 2.0.0
+
+* Add ability to iterate from the right instead of the left
+
+#### BREAKING CHANGES
+
+* All options after initial value are now an object of options
+
+```javascript
+// before
+transduce(map(fn), collection, initialValue, passHandler);
+
+// now
+transduce(map(fn), collection, initialValue, { passHandler });
+```
+
 ## 1.2.0
 
 * Add [`find`](README.md#find) method